Render Markdown in task header
