+++ webview-ui/src/components/common/MarkdownBlock.styles.ts
@@ -1,0 +1,65 @@
+import styled from "styled-components"
+
+export const StyledMarkdown = styled.div`
+	code:not(pre > code) {
+		font-family: var(--vscode-editor-font-family, monospace);
+		color: var(--vscode-textPreformat-foreground, #f78383);
+		background-color: var(--vscode-textCodeBlock-background, #1e1e1e);
+		padding: 0px 2px;
+		border-radius: 3px;
+		border: 1px solid var(--vscode-textSeparator-foreground, #424242);
+		white-space: pre-line;
+		word-break: break-word;
+		overflow-wrap: anywhere;
+	}
+
+	/* Target only Dark High Contrast theme using the data attribute VS Code adds to the body */
+	body[data-vscode-theme-kind="vscode-high-contrast"] & code:not(pre > code) {
+		color: var(
+			--vscode-editorInlayHint-foreground,
+			var(--vscode-symbolIcon-stringForeground, var(--vscode-charts-orange, #e9a700))
+		);
+	}
+
+	font-family:
+		var(--vscode-font-family),
+		system-ui,
+		-apple-system,
+		BlinkMacSystemFont,
+		"Segoe UI",
+		Roboto,
+		Oxygen,
+		Ubuntu,
+		Cantarell,
+		"Open Sans",
+		"Helvetica Neue",
+		sans-serif;
+
+	font-size: var(--vscode-font-size, 13px);
+
+	p,
+	li,
+	ol,
+	ul {
+		line-height: 1.25;
+	}
+
+	ol,
+	ul {
+		padding-left: 2.5em;
+		margin-left: 0;
+	}
+
+	p {
+		white-space: pre-wrap;
+		margin-top: 0;
+		margin-bottom: 0.5em;
+	}
+
+	a {
+		text-decoration: none;
+		&:hover {
+			text-decoration: underline;
+		}
+	}
+`
