+++ webview-ui/src/components/common/MarkdownBlock.tsx
@@ -1,192 +1,50 @@
 import React, { memo, useEffect } from "react"
 import { useRemark } from "react-remark"
-import styled from "styled-components"
-import { visit } from "unist-util-visit"
+import { ComponentOptions } from "rehype-react"
 
+import { remarkUrls, remarkCodeLang } from "@src/utils/remark"
+import { rehypeMentions, rehypeCode, rehypeMermaid } from "@src/utils/rehype"
 import { useExtensionState } from "@src/context/ExtensionStateContext"
 
-import CodeBlock from "./CodeBlock"
-import MermaidBlock from "./MermaidBlock"
+import { StyledMarkdown } from "./MarkdownBlock.styles"
+
+export type RehypePlugin = "mentions" | "code" | "mermaid"
 
 interface MarkdownBlockProps {
 	markdown?: string
+	renderMentions?: boolean
+	renderCode?: boolean
+	renderMermaid?: boolean
 }
 
-/**
- * Custom remark plugin that converts plain URLs in text into clickable links
- *
- * The original bug: We were converting text nodes into paragraph nodes,
- * which broke the markdown structure because text nodes should remain as text nodes
- * within their parent elements (like paragraphs, list items, etc.).
- * This caused the entire content to disappear because the structure became invalid.
- */
-const remarkUrlToLink = () => {
-	return (tree: any) => {
-		// Visit all "text" nodes in the markdown AST (Abstract Syntax Tree)
-		visit(tree, "text", (node: any, index, parent) => {
-			const urlRegex = /https?:\/\/[^\s<>)"]+/g
-			const matches = node.value.match(urlRegex)
+export const MarkdownBlock = memo(
+	({ markdown, renderMentions = false, renderCode = true, renderMermaid = true }: MarkdownBlockProps) => {
+		const { theme } = useExtensionState()
 
-			if (!matches) {
-				return
-			}
+		const components: ComponentOptions<typeof React.createElement>["components"] = {}
 
-			const parts = node.value.split(urlRegex)
-			const children: any[] = []
-
-			parts.forEach((part: string, i: number) => {
-				if (part) {
-					children.push({ type: "text", value: part })
-				}
-
-				if (matches[i]) {
-					children.push({ type: "link", url: matches[i], children: [{ type: "text", value: matches[i] }] })
-				}
-			})
-
-			// Fix: Instead of converting the node to a paragraph (which broke things),
-			// we replace the original text node with our new nodes in the parent's children array.
-			// This preserves the document structure while adding our links.
-			if (parent) {
-				parent.children.splice(index, 1, ...children)
-			}
-		})
-	}
-}
-
-const StyledMarkdown = styled.div`
-	code:not(pre > code) {
-		font-family: var(--vscode-editor-font-family, monospace);
-		color: var(--vscode-textPreformat-foreground, #f78383);
-		background-color: var(--vscode-textCodeBlock-background, #1e1e1e);
-		padding: 0px 2px;
-		border-radius: 3px;
-		border: 1px solid var(--vscode-textSeparator-foreground, #424242);
-		white-space: pre-line;
-		word-break: break-word;
-		overflow-wrap: anywhere;
-	}
-
-	/* Target only Dark High Contrast theme using the data attribute VS Code adds to the body */
-	body[data-vscode-theme-kind="vscode-high-contrast"] & code:not(pre > code) {
-		color: var(
-			--vscode-editorInlayHint-foreground,
-			var(--vscode-symbolIcon-stringForeground, var(--vscode-charts-orange, #e9a700))
-		);
-	}
-
-	font-family:
-		var(--vscode-font-family),
-		system-ui,
-		-apple-system,
-		BlinkMacSystemFont,
-		"Segoe UI",
-		Roboto,
-		Oxygen,
-		Ubuntu,
-		Cantarell,
-		"Open Sans",
-		"Helvetica Neue",
-		sans-serif;
-
-	font-size: var(--vscode-font-size, 13px);
-
-	p,
-	li,
-	ol,
-	ul {
-		line-height: 1.25;
-	}
-
-	ol,
-	ul {
-		padding-left: 2.5em;
-		margin-left: 0;
-	}
-
-	p {
-		white-space: pre-wrap;
-	}
-
-	a {
-		text-decoration: none;
-	}
-	a {
-		&:hover {
-			text-decoration: underline;
+		if (renderMentions) {
+			components.p = rehypeMentions
 		}
-	}
-`
 
-const MarkdownBlock = memo(({ markdown }: MarkdownBlockProps) => {
-	const { theme } = useExtensionState()
-	const [reactContent, setMarkdown] = useRemark({
-		remarkPlugins: [
-			remarkUrlToLink,
-			() => {
-				return (tree) => {
-					visit(tree, "code", (node: any) => {
-						if (!node.lang) {
-							node.lang = "text"
-						} else if (node.lang.includes(".")) {
-							node.lang = node.lang.split(".").slice(-1)[0]
-						}
-					})
-				}
-			},
-		],
-		rehypePlugins: [],
-		rehypeReactOptions: {
-			components: {
-				pre: ({ node: _, children }: any) => {
-					// Check for Mermaid diagrams first
-					if (Array.isArray(children) && children.length === 1 && React.isValidElement(children[0])) {
-						const child = children[0] as React.ReactElement<{ className?: string }>
+		if (renderCode) {
+			components.pre = rehypeCode
+		}
 
-						if (child.props?.className?.includes("language-mermaid")) {
-							return child
-						}
-					}
+		if (renderMermaid) {
+			components.code = rehypeMermaid
+		}
 
-					// For all other code blocks, use CodeBlock with copy button
-					const codeNode = children?.[0]
+		const [reactContent, setMarkdown] = useRemark({
+			remarkPlugins: [remarkUrls, remarkCodeLang],
+			rehypePlugins: [],
+			rehypeReactOptions: { components },
+		})
 
-					if (!codeNode?.props?.children) {
-						return null
-					}
+		useEffect(() => setMarkdown(markdown || ""), [markdown, setMarkdown, theme])
 
-					const language =
-						(Array.isArray(codeNode.props?.className)
-							? codeNode.props.className
-							: [codeNode.props?.className]
-						).map((c: string) => c?.replace("language-", ""))[0] || "javascript"
-
-					const rawText = codeNode.props.children[0] || ""
-					return <CodeBlock source={rawText} language={language} />
-				},
-				code: (props: any) => {
-					const className = props.className || ""
-
-					if (className.includes("language-mermaid")) {
-						const codeText = String(props.children || "")
-						return <MermaidBlock code={codeText} />
-					}
-
-					return <code {...props} />
-				},
-			},
-		},
-	})
-
-	useEffect(() => {
-		setMarkdown(markdown || "")
-	}, [markdown, setMarkdown, theme])
-
-	return (
-		<div style={{}}>
-			<StyledMarkdown>{reactContent}</StyledMarkdown>
-		</div>
-	)
-})
+		return <StyledMarkdown>{reactContent}</StyledMarkdown>
+	},
+)
 
 export default MarkdownBlock
