+++ webview-ui/src/components/common/MermaidBlock.tsx
@@ -1,8 +1,11 @@
 import { useEffect, useRef, useState } from "react"
 import mermaid from "mermaid"
+
 import { useDebounceEffect } from "@src/utils/useDebounceEffect"
-import styled from "styled-components"
 import { vscode } from "@src/utils/vscode"
+import { svgToPng } from "@src/utils/svgToPng"
+
+import { MermaidBlockContainer, LoadingMessage, SvgContainer } from "./MermaidBlock.styles"
 
 const MERMAID_THEME = {
 	background: "#1e1e1e", // VS Code dark theme background
@@ -124,16 +127,19 @@ export default function MermaidBlock({ code }: MermaidBlockProps) {
 	 * Converts the <svg> to a PNG and sends it to the extension.
 	 */
 	const handleClick = async () => {
-		if (!containerRef.current) return
+		if (!containerRef.current) {
+			return
+		}
+
 		const svgEl = containerRef.current.querySelector("svg")
-		if (!svgEl) return
+
+		if (!svgEl) {
+			return
+		}
 
 		try {
-			const pngDataUrl = await svgToPng(svgEl)
-			vscode.postMessage({
-				type: "openImage",
-				text: pngDataUrl,
-			})
+			const pngDataUrl = await svgToPng(svgEl, MERMAID_THEME)
+			vscode.postMessage({ type: "openImage", text: pngDataUrl })
 		} catch (err) {
 			console.error("Error converting SVG to PNG:", err)
 		}
@@ -142,85 +148,7 @@ export default function MermaidBlock({ code }: MermaidBlockProps) {
 	return (
 		<MermaidBlockContainer>
 			{isLoading && <LoadingMessage>Generating mermaid diagram...</LoadingMessage>}
-
-			{/* The container for the final <svg> or raw code. */}
 			<SvgContainer onClick={handleClick} ref={containerRef} $isLoading={isLoading} />
 		</MermaidBlockContainer>
 	)
 }
-
-async function svgToPng(svgEl: SVGElement): Promise<string> {
-	// Clone the SVG to avoid modifying the original
-	const svgClone = svgEl.cloneNode(true) as SVGElement
-
-	// Get the original viewBox
-	const viewBox = svgClone.getAttribute("viewBox")?.split(" ").map(Number) || []
-	const originalWidth = viewBox[2] || svgClone.clientWidth
-	const originalHeight = viewBox[3] || svgClone.clientHeight
-
-	// Calculate the scale factor to fit editor width while maintaining aspect ratio
-
-	// Unless we can find a way to get the actual editor window dimensions through the VS Code API (which might be possible but would require changes to the extension side),
-	// the fixed width seems like a reliable approach.
-	const editorWidth = 3_600
-
-	const scale = editorWidth / originalWidth
-	const scaledHeight = originalHeight * scale
-
-	// Update SVG dimensions
-	svgClone.setAttribute("width", `${editorWidth}`)
-	svgClone.setAttribute("height", `${scaledHeight}`)
-
-	const serializer = new XMLSerializer()
-	const svgString = serializer.serializeToString(svgClone)
-	const svgDataUrl = "data:image/svg+xml;base64," + btoa(decodeURIComponent(encodeURIComponent(svgString)))
-
-	return new Promise((resolve, reject) => {
-		const img = new Image()
-		img.onload = () => {
-			const canvas = document.createElement("canvas")
-			canvas.width = editorWidth
-			canvas.height = scaledHeight
-
-			const ctx = canvas.getContext("2d")
-			if (!ctx) return reject("Canvas context not available")
-
-			// Fill background with Mermaid's dark theme background color
-			ctx.fillStyle = MERMAID_THEME.background
-			ctx.fillRect(0, 0, canvas.width, canvas.height)
-
-			ctx.imageSmoothingEnabled = true
-			ctx.imageSmoothingQuality = "high"
-
-			ctx.drawImage(img, 0, 0, editorWidth, scaledHeight)
-			resolve(canvas.toDataURL("image/png", 1.0))
-		}
-		img.onerror = reject
-		img.src = svgDataUrl
-	})
-}
-
-const MermaidBlockContainer = styled.div`
-	position: relative;
-	margin: 8px 0;
-`
-
-const LoadingMessage = styled.div`
-	padding: 8px 0;
-	color: var(--vscode-descriptionForeground);
-	font-style: italic;
-	font-size: 0.9em;
-`
-
-interface SvgContainerProps {
-	$isLoading: boolean
-}
-
-const SvgContainer = styled.div<SvgContainerProps>`
-	opacity: ${(props) => (props.$isLoading ? 0.3 : 1)};
-	min-height: 20px;
-	transition: opacity 0.2s ease;
-	cursor: pointer;
-	display: flex;
-	justify-content: center;
-`
